WeatherEmail/requirements.txt
certifi==2019.6.16
chardet==3.0.4
DateTime==4.3
Django==2.2.2
idna==2.8
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
zope.interface==4.6.0
